requirements.txt
langchain-openai
langchain-agents
langchain-core
langchain-community